Some changes

+++ src/private/mt_run_student.m
@@ -7,33 +7,42 @@ function mt_run_student(student_data, mark, overall_remarks, ...
                         questions_title, questions_remarks, mt_settings)
 student_filepath = mt_filepath(mt_settings.output_dir, student_data);
 
-%% Generate latex file and write it out
-latex = mt_generate_tex(student_data, mark, overall_remarks, ...
-                        questions_title, questions_remarks, ...
-                        mt_settings);
-mt_write_text_file(student_filepath + '.tex', latex);
 
-%% Create pdf file
-command_compile = char(strcat(mt_settings.pdflatex_filename, {' '}, ...
-             {'-output-directory '}, mt_settings.output_dir, {' '}, ...
-             student_filepath, '.tex'));
-system(command_compile);
-system(command_compile); % Render references
+if mt_settings.generate_pdf
+    % Generate latex file and write it out
+    latex = mt_generate_tex(student_data, mark, overall_remarks, ...
+                            questions_title, questions_remarks, ...
+                            mt_settings);
+    mt_write_text_file(student_filepath + '.tex', latex);
 
-%% Delete temporary files
-system(char(strcat({'rm '}, student_filepath, '.aux')));
-system(char(strcat({'rm '}, student_filepath, '.log')));
-system(char(strcat({'rm '}, student_filepath, '.out')));
+    % Create pdf file
+    command_compile = char(strcat(mt_settings.pdflatex_filename, {' '}, ...
+                 {'-output-directory '}, mt_settings.output_dir, {' '}, ...
+                 student_filepath, '.tex'));
+    system(command_compile);
+    system(command_compile); % Render references
+
+    % Delete temporary files
+    system(char(strcat({'rm '}, student_filepath, '.aux')));
+    system(char(strcat({'rm '}, student_filepath, '.log')));
+    system(char(strcat({'rm '}, student_filepath, '.out')));
+end
 
 %% Create email
 if mt_settings.send_emails
     email = mt_generate_email(student_data, mark, mt_settings);
     mt_write_text_file(student_filepath + '.txt', email);
     
+    if mt_settings.generate_pdf
+        attachment = strcat(pwd, '/', student_filepath, '.pdf');
+    else
+        attachment = '';
+    end
+    
     script = mt_generate_applescript(student_data.email, ...
                                      mt_settings.email_subject, ...
                                      email, ...
-                                     strcat(pwd, '/', student_filepath, '.pdf'), ... 
+                                     attachment, ... 
                                      mt_settings);
     mt_write_text_file(student_filepath + '.script', script);
 end
